Optimize _CC definition
Optimize _CC definition to prevent users from doing compilation-time
operation in _CC{}CC_ blocks by hiding __this_function with a local char
variable with the same name. In this way, the misuse of instruction
calls in run-time can be detected by C++ compiler as long as _DEBUG or
ASYS_DEBUG is defined.